+++ UniEngine/DirectionalLightShadowMap.h
@@ -11,7 +11,7 @@ namespace UniEngine {
 		size_t _LightAmount;
 		void Allocate();
 	public:
-		DirectionalLightShadowMap(size_t amount, float resolutionX = 1024.0f, float resolutionY = 1024.0f, bool enableVSM = false);
+		DirectionalLightShadowMap(size_t amount, size_t resolutionX = 1024, size_t resolutionY = 1024, bool enableVSM = false);
 		void SetVSM(bool enabled);
 		void SetLightAmount(size_t value);
 		GLTexture2DArray* DepthMapArray();
